Change to tsfsmooth to prevent compiler warning

+++ cmd/tsfsmooth.cpp
@@ -74,7 +74,7 @@ void run ()
   while (reader (tck_scalar)) {
     std::vector<value_type> tck_scalars_smoothed (tck_scalar.size());
 
-    for (size_t i = 0; i < tck_scalar.size(); ++i) {
+    for (int i = 0; i < (int)tck_scalar.size(); ++i) {
       float norm_factor = 0.0;
       float value = 0.0;
       for (int k = -(int)radius; k <= (int)radius; ++k) {
